+++ tests/test_enterprise/api/test_views.py
@@ -8809,6 +8809,36 @@ class TestEnterpriseGroupViewSet(APITest):
         assert pending_membership.status == GROUP_MEMBERSHIP_PENDING_STATUS
         assert membership.status == GROUP_MEMBERSHIP_ACCEPTED_STATUS
 
+    def test_assign_learners_case_insensitive_emails(self):
+        """
+        Test assigning learners to a group with emails whose case differs from their registered user emails
+        """
+        new_group = EnterpriseGroupFactory(enterprise_customer=self.enterprise_customer)
+        user_1 = UserFactory(email='testuser1@example.com')
+        user_2 = UserFactory(email='TestUser2@example.com')
+        EnterpriseCustomerUserFactory(enterprise_customer=self.enterprise_customer, user_id=user_1.id)
+        EnterpriseCustomerUserFactory(enterprise_customer=self.enterprise_customer, user_id=user_2.id)
+
+        assign_url = settings.TEST_SERVER + reverse(
+            'enterprise-group-assign-learners',
+            kwargs={'group_uuid': new_group.uuid},
+        )
+        request_data = {
+            'learner_emails': ['TestUser1@example.com', 'testuser2@example.com'],
+        }
+        self.client.post(assign_url, data=request_data)
+
+        # Verify both group memberships in place
+        assert len(EnterpriseGroupMembership.objects.filter(
+            group=new_group,
+            pending_enterprise_customer_user__isnull=True)
+        ) == 2
+        # Verify no pending group memberships
+        assert len(EnterpriseGroupMembership.objects.filter(
+            group=new_group,
+            pending_enterprise_customer_user__isnull=False)
+        ) == 0
+
     @mock.patch('enterprise.tasks.send_group_membership_invitation_notification.delay', return_value=mock.MagicMock())
     def test_successful_assign_learners_to_group(self, mock_send_group_membership_invitation_notification):
         """